+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Changes:
 
 - Extend support for `WrapperKeepOpaque` with internal decoding (allowed to fail)
 - Adjust max-entries in historic provider LRU
+- Update to latest Substrate/Polkadot/Kusama metadata
 - Update `@polkadot/util` to 8.3.1
 - Update package paths under ESM